image-segmentation/requirements.txt
roboflow
ultralytics
python-dotenv
torch
numpy
Pillow
opencv-python
shapely
pywebview